finish conduit bandwidth test program.
figure the bug about using async_read/write operation. It's because when
a 'close' signal send to extra thread by async-finish function, before
the extra-thread notice this flag, there may be new msg add to the
queue, however, the 'close' signal is till set. So we need reset 'close'
signal to false when we add new item to queue.

Basically right now the inproc/cross-proc conduit read/write, including
block and nonblock, buffered or pointered, operations are completeed.
But the performance is kind of bad, especially for inproc conduit. This
is mainly due to the highly using of mutex and condition var to do the
lock/unlock ops. 

Next version, based on current task and conduit frame, optimize conduit
implementation. Try to use lock-free algorithm to redesign the conduit
implementation.