Change task init/run/waittilldone call end execute design.
Now after task-threads created, thread will do a while(1) loop to check
jobqueue, task.init and other function will add a job to queue. 

Also add a new task job 'finish' which will make thread to jump out of
while(1) and finish thread. This function will be call by task
destructor automatically, or user can call it directly, but can only be
called once.

There's a problem about conduit when it destruct too early. Before we
deal with this by checking and waiting task-threads finish, but this is
not a good idear, we should not compel this behavier. We can deal with
this problem by construct conduit always as a pointer, and it's newed,
then we could destruct conduit in conduit manager.