Finish debug r/w byfirst operation.
The opbyfirst and unique exection now works, noticing how to reuse the
atomic vars related with these function.

Also we adopt the same solve method to multi-threads op control. The
preveous unlimited vector problem is solved, but it's based on the
assumpution that async operation between different threads usually are
not very long, means that there will not be case: many many r/w ops
happens, and no synch betweed threads, if this happen, a thread may
stuck at very early, and another thread goes at the front, and may reuse
the atomic or latch that the early threads still not finished using. And
this would cause programm undefind behaver and error.
Be careful of this bug!!!