Finish debug bcast and gather.
Finish debug k-means of using bast and gather in task.

Add '#define SET_CPU_AFFINITY', if set this, each thead in a proc will
be bind to a unique cpu core. When total threads of all task in a proc
is less than TOTAL HARD CORES, this would be better performance, but
else, sync and send/recv between thread that bind to same core would
cause big latency. Then prefered to use 'bind-to-none'=='bind-to-all'.