Redo the latency test.
Find out why the latecy increase unevenly when the msg size double. It's
related to our "wait" methods. For big msg, we pass by ptr, so the
writer will wait for reader to finish. The wait we use is a busy wait,
and use "pause"  "yield" "sleep" for it. The latency jump is caused by
change from yield to sleep, so we may need to increase the number of
using "pause" and "yield",  sleep may cost more delay than we expected.