Complete the heat2D and heat image and kmeans test.
Add a setUserTaskObj() in int Task class. But still do more.
as we create the real task artifact(CPU/GPU) when define the task
object. So, when later we set a new user class template object for
using, we also need update this objest to real task artifact, else we
still use the old user task object for init/run method call. As we only
pass the user task objet pointer to task artifact, so when we user a new
one, the other pointer still point to the old one.