Complete scoped_shmem related functions.
It depends on the underlying internal_MPIWin class. 
each scoped_shmem function will need a internal_MPIWin obj as the
arguments.

The shmem-collective operations are not implemented, as they are not
related to one-side comm, will be implemented in other places.

The correction of lock()/unlock() of MpiWinLock class is not sure, may
be problematic!!! need further explore.

Next work: 
		1. implement old "global_scoped_data" on top of this "scoped_shmem" 
		2. compile, debug, and test.