Add GlobalGpuData class.
For global shared data with GPU, we create a separate GlobalGpuData class to implemente it. For simplicity, we allocate memory by cuda_malloc_manage(cuda unified memory), and create a mpi window with this memory.
provide two simple rload/rstore method to get/put remote entire data.  This window use the same communicator as GlobalScoped data object for a task, but each instance of this class will ceate a new window. (in Global Scoped data, we only has one window for a task, and use mspace_malloc to use the memory in this window).

Change the make file for the project, now each group of sorce files use a Make.mk for making objs, and for a platform there is a .comm file define common info of making procedure.