Go on debug global data test.
Now the test can run. Becare full the internal base task object, which are created on every process. But the user task object, we only created on those required process, so care some data structure are invalid on some process for a task, and the problem when destruct a user task object and base task object.