The global-scoped-test works ok.
Add both taskgroupProcess rank -> world process rank map and  worldrank->group rank.
When use global scoped data, and in a task implementation code, usually we just use the process's group rank, and ops in task's related communicator.

Now the process rank used in remote load/store ops are group-rank, we will not transit it.

For mpi program, std::in is only forwarded to process rank 0, not work for other processes, so , for input data, usually we need use a file to read for when programs is running.

Sometimes when we change some source code file(.c or .h or .inc) we may need to rebuild the whole objs, otherwise some objs are old and case error.