Finish rotate-cpu version
As for rotate/rgbyub/cray/bfs, meaningless to do on mutiplenodes. For multinodes, we will create a batch version.

When use global-shared data object, do not access to ileagle address, it may not post error, but when do global.destroy() it may post error.

add messge rate(as well can be seen as bandwidth) test for rstore operation.

add mpi bandwidth using multiple pairs of threads.