finish MM utc-cpu version.
For mm, using rload/store is not as good as use broadcast /gather.

Notice when call task.run multiple times, later one should be called after call wait for the former one.

As we use args to pass some info to task object, becareful if the args is valid on each process. Most cases, we only assume the args is valid on 0 proc. So, in main program, we may need check current proc rank to do nessaray job.
This is a main drawback for current implementation, as we use mpi platform to init the execution on remote node. It's kind of symetric (SPMD).  Ideally, the main program should play as the master, invoke and districute tasks to their required nodes for running. In each task code, we will treat it as an SPMD program.