Noticing the new cpu bind feature.
Seems for some program, bind threads of a task to same CPU will reduce intra-sync overhead, but most cases,  choose bind to cpu is a better option, and use  all hyper threads of that CPU.

Becareful the HC test. Kind of strange that the performance of this program is highly sensitive to a data write operation in the kernel computing procedure.  And also cause large sync overhead than other test programs.